Improve game detection in "Open Map" modal dialog
This affects opening a PWAD for the first time when the .dbs file
is missing. We can better detect the game config to use by analyzing
naming formats of map lumps.

If MAPxy is found in the given WAD, don't auto-select:
* Doom 1
* Heretic

If ExMy is found, don't auto-select:
* Doom 2
* Hexen
* Strife

So the first-matching item in the game configuration dropdown won't
be an obvious mismatch.

This doesn't affect Eternity Engine or ZDoom-derivative configs that
support arbitrary map names.